Relocate task intrinsic tests to integration
